Move the exit utils into a separate file
